requirements.txt
yfinance==0.2.63
pandas==2.3.0
numpy==2.3.0
requests==2.32.4
textblob==0.19.0
python-dotenv==1.0.0 